+++ src/core/energy_storage/thermal_storage.jl
@@ -22,6 +22,8 @@ Cold thermal energy storage sytem; specifically, a chilled water system used to 
     macrs_itc_reduction::Float64 = 0.5
     total_itc_fraction::Float64 = 0.3
     total_rebate_per_kwh::Float64 = 0.0
+    min_duration_hours::Real = 0.0
+    max_duration_hours::Real = 100000.0
 ```
 """
 Base.@kwdef struct ColdThermalStorageDefaults <: AbstractThermalStorageDefaults
@@ -40,6 +42,8 @@ Base.@kwdef struct ColdThermalStorageDefaults <: AbstractThermalStorageDefaults
     macrs_itc_reduction::Float64 = 0.5
     total_itc_fraction::Float64 = 0.3
     total_rebate_per_kwh::Float64 = 0.0
+    min_duration_hours::Real = 0.0
+    max_duration_hours::Real = 100000.0
 end
 
 
@@ -65,6 +69,8 @@ end
     can_serve_dhw::Bool = true
     can_serve_space_heating:Bool = true
     can_serve_process_heat::Bool = false
+    min_duration_hours::Real = 0.0
+    max_duration_hours::Real = 100000.0
 ```
 """
 Base.@kwdef struct HotThermalStorageDefaults <: AbstractThermalStorageDefaults
@@ -86,6 +92,8 @@ Base.@kwdef struct HotThermalStorageDefaults <: AbstractThermalStorageDefaults
     can_serve_dhw::Bool = true
     can_serve_space_heating::Bool = true
     can_serve_process_heat::Bool = false
+    min_duration_hours::Real = 0.0
+    max_duration_hours::Real = 100000.0
 end
 
 
@@ -119,6 +127,8 @@ struct ColdThermalStorage <: AbstractThermalStorage
     discharge_efficiency::Float64
     net_present_cost_per_kwh::Float64
     om_cost_per_kwh::Float64
+    min_duration_hours::Real
+    max_duration_hours::Real
 
     function ColdThermalStorage(s::AbstractThermalStorageDefaults, f::Financial, time_steps_per_hour::Int)
          
@@ -167,7 +177,9 @@ struct ColdThermalStorage <: AbstractThermalStorage
             charge_efficiency,
             discharge_efficiency,
             net_present_cost_per_kwh,
-            om_cost_per_kwh
+            om_cost_per_kwh,
+            s.min_duration_hours,
+            s.max_duration_hours
         )
     end
 end
@@ -205,6 +217,9 @@ struct HotThermalStorage <: AbstractThermalStorage
     can_serve_dhw::Bool
     can_serve_space_heating::Bool
     can_serve_process_heat::Bool
+    min_duration_hours::Real
+    max_duration_hours::Real
+    
 
     function HotThermalStorage(s::AbstractThermalStorageDefaults, f::Financial, time_steps_per_hour::Int)
          
@@ -256,7 +271,9 @@ struct HotThermalStorage <: AbstractThermalStorage
             om_cost_per_kwh,
             s.can_serve_dhw,
             s.can_serve_space_heating,
-            s.can_serve_process_heat
+            s.can_serve_process_heat,
+            s.min_duration_hours,
+            s.max_duration_hours
         )
     end
 end